+++ src/modules/infra/middlewares/hocs/base.ts
@@ -1,3 +1,0 @@
-export { defineGlobalInterceptorHOC } from './interceptors.ts'
-export { defineGlobalPipeHOC } from './pipes.ts'
-export { defineGlobalGuardHOC } from './guards.ts'